Move metric test utils out of go.vumitools.tests.test_metrics and into go.vumitools.test.utils
